backend/pyproject.toml
[project]
name = "image-search-v2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cors>=1.0.1",
    "flask>=3.1.1",
    "flask-cors>=6.0.1",
    "open-clip-torch>=2.32.0",
    "pillow>=11.3.0",
    "python-dotenv>=1.1.1",
    "qdrant-client>=1.14.3",
    "torch>=2.7.1",
]
